+++ dataSource/repositoryAdapters/src/test/java/pl/zzpj/repository/mappers/TransactionMapperTest.java
@@ -10,6 +10,7 @@ import pl.zzpj.model.Account;
 import pl.zzpj.model.Currency;
 import pl.zzpj.model.Transaction;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 import java.time.Instant;
 
@@ -32,14 +33,14 @@ class TransactionMapperTest {
         Timestamp timestamp = Timestamp.from(Instant.now());
 
         Transaction transaction = new Transaction();
-        transaction.setAmount(100);
+        transaction.setAmount(BigDecimal.valueOf(100));
         transaction.setDate(timestamp);
         transaction.setFrom(from);
         transaction.setTo(to);
 
         TransactionEnt transactionEnt = TransactionMapper.mapToTransactionEnt(transaction);
 
-        assertEquals(100, transactionEnt.getAmount());
+        assertEquals(BigDecimal.valueOf(100), transactionEnt.getAmount());
         assertEquals(timestamp, transactionEnt.getDate());
         assertEquals(from.getLogin(), transactionEnt.getFrom().getLogin());
         assertEquals(to.getLogin(), transactionEnt.getTo().getLogin());
@@ -60,14 +61,14 @@ class TransactionMapperTest {
         Timestamp timestamp = Timestamp.from(Instant.now());
 
         TransactionEnt transactionEnt = new TransactionEnt();
-        transactionEnt.setAmount(100);
+        transactionEnt.setAmount(BigDecimal.valueOf(100));
         transactionEnt.setDate(timestamp);
         transactionEnt.setFrom(from);
         transactionEnt.setTo(to);
 
         Transaction transaction = TransactionMapper.mapToTransaction(transactionEnt);
 
-        assertEquals(100, transaction.getAmount());
+        assertEquals(BigDecimal.valueOf(100), transaction.getAmount());
         assertEquals(timestamp, transaction.getDate());
         assertEquals(from.getLogin(), transaction.getFrom().getLogin());
         assertEquals(to.getLogin(), transaction.getTo().getLogin());
